# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Edubee Camp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: campProviderId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractListResponse"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a new contract
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contract found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
    put:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "200":
          description: Contract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /invoices:
    get:
      operationId: listInvoices
      tags: [finance]
      summary: List all invoices
      security:
        - BearerAuth: []
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: invoiceType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [finance]
      summary: Create a new invoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [finance]
      summary: Get invoice by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [finance]
      summary: Update invoice
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List all transactions
      security:
        - BearerAuth: []
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
        - name: invoiceId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create a new transaction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /notifications:
    get:
      operationId: listNotifications
      tags: [reports]
      summary: List notifications for current user
      security:
        - BearerAuth: []
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [reports]
      summary: Mark notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [reports]
      summary: Get dashboard statistics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /exchange-rates:
    get:
      operationId: listExchangeRates
      tags: [finance]
      summary: List exchange rates
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateListResponse"
    post:
      operationId: createExchangeRate
      tags: [finance]
      summary: Create or update exchange rate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExchangeRateRequest"
      responses:
        "201":
          description: Exchange rate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRate"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    Application:
      type: object
      properties:
        id:
          type: string
        applicationNumber:
          type: string
          nullable: true
        agentId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        packageGroupId:
          type: string
          nullable: true
        packageId:
          type: string
          nullable: true
        preferredStartDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
        totalChildren:
          type: integer
        totalAdults:
          type: integer
        primaryLanguage:
          type: string
        referralSource:
          type: string
          nullable: true
        specialRequests:
          type: string
          nullable: true
        termsAccepted:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - totalChildren
        - totalAdults
        - primaryLanguage
        - termsAccepted
        - createdAt
        - updatedAt
...
    Contract:
      type: object
      properties:
        id:
          type: string
        contractNumber:
          type: string
          nullable: true
        applicationId:
          type: string
          nullable: true
        campProviderId:
          type: string
          nullable: true
        totalAmount:
          type: string
          nullable: true
        currency:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - currency
        - status
        - createdAt
        - updatedAt
    ContractDetail:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            application:
              $ref: "#/components/schemas/Application"
              nullable: true
    CreateContractRequest:
      type: object
      properties:
        applicationId:
          type: string
          nullable: true
        campProviderId:
          type: string
          nullable: true
        totalAmount:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    ContractListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        contractId:
          type: string
          nullable: true
        invoiceType:
          type: string
          enum:
            - client
            - agent
            - partner
        recipientId:
          type: string
          nullable: true
        subtotal:
          type: string
          nullable: true
        taxAmount:
          type: string
          nullable: true
        totalAmount:
          type: string
          nullable: true
        currency:
          type: string
        status:
          type: string
        issuedAt:
          type: string
          format: date-time
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceType
        - currency
        - status
        - createdAt
        - updatedAt
    CreateInvoiceRequest:
      type: object
      properties:
        contractId:
          type: string
          nullable: true
        invoiceType:
          type: string
        recipientId:
          type: string
          nullable: true
        lineItems:
          type: object
          nullable: true
          additionalProperties: true
        subtotal:
          type: string
          nullable: true
        taxAmount:
          type: string
          nullable: true
        totalAmount:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - invoiceType
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Transaction:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
          nullable: true
        invoiceId:
          type: string
          nullable: true
        transactionType:
          type: string
        amount:
          type: string
          nullable: true
        currency:
          type: string
        description:
          type: string
          nullable: true
        bankReference:
          type: string
          nullable: true
        transactionDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transactionType
        - currency
        - createdAt
    CreateTransactionRequest:
      type: object
      properties:
        contractId:
          type: string
          nullable: true
        invoiceId:
          type: string
          nullable: true
        transactionType:
          type: string
        amount:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        bankReference:
          type: string
          nullable: true
        transactionDate:
          type: string
          format: date
          nullable: true
      required:
        - transactionType
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        referenceType:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - isRead
        - createdAt
    NotificationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    DashboardStats:
      type: object
      properties:
        totalApplications:
          type: integer
        pendingApplications:
          type: integer
        activeContracts:
          type: integer
        totalRevenue:
          type: string
        totalLeads:
          type: integer
        newLeadsThisMonth:
          type: integer
        recentApplications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        applicationsByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required:
              - status
              - count
      required:
        - totalApplications
        - pendingApplications
        - activeContracts
        - totalRevenue
        - totalLeads
        - newLeadsThisMonth
        - recentApplications
        - applicationsByStatus
    ExchangeRate:
      type: object
      properties:
        id:
          type: string
        fromCurrency:
          type: string
        toCurrency:
          type: string
        rate:
          type: string
        source:
          type: string
          nullable: true
        effectiveDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fromCurrency
        - toCurrency
        - rate
        - effectiveDate
        - createdAt
    CreateExchangeRateRequest:
      type: object
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
          nullable: true
        rate:
          type: string
        source:
          type: string
          nullable: true
        effectiveDate:
          type: string
          format: date
      required:
        - fromCurrency
        - rate
        - effectiveDate
    ExchangeRateListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeRate"
      required:
        - data