# lib/api-zod/src/generated/types/createContractRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Edubee Camp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractRequest {
  applicationId?: string | null;
  campProviderId?: string | null;
  totalAmount?: string | null;
  currency?: string | null;
  status?: string | null;
  startDate?: Date | null;
  endDate?: Date | null;
  notes?: string | null;
}