# lib/api-zod/src/generated/types/createExchangeRateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Edubee Camp API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExchangeRateRequest {
  fromCurrency: string;
  toCurrency?: string | null;
  rate: string;
  source?: string | null;
  effectiveDate: Date;
}